If-statement for channel delete not returning properly + wrong data type in channel mapping

+++ rockpack/mainsite/core/es/mappings.py
@@ -130,7 +130,7 @@ channel_mapping = {
                     "index": "not_analyzed"
                 },
                 "aoi": {
-                    "type": "integer",
+                    "type": "float",
                 }
             },
         },
